requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
Pillow==10.0.1
flask-socketio==5.3.6
python-socketio==5.8.0
numpy==1.24.3
firebase-admin==6.2.0
python-jose[cryptography]==3.3.0
firebase-admin==7.2.0
pyrebase4==4.9.0
tensorflow==2.15.0
scikit-learn==1.3.2
requests==2.32.5
